Created GridView component, started to use it in forms index and steps index.

+++ resources/js/store/modules/forms.js
@@ -78,6 +78,10 @@ export default {
             }
         },
         deleteForm({commit}, {id, index}) {
+            if (!confirm('Do you really want to delete this item')) {
+                return;
+            }
+
             return new Promise((resolve, reject) => {
                 forms.deleteForm(
                     response => {
